Silent PR to Master to check updated documentation

+++ tests/test_sir_estimate.py
@@ -27,7 +27,7 @@ class TestModelEstimate(TestCase):
         self.target = np.array([0.5, 1.0/3.0])
 
         # constraints
-        EPSILON = np.sqrt(np.finfo(np.float).eps)
+        EPSILON = np.sqrt(np.finfo(float).eps)
 
         self.box_bounds = [(EPSILON, 5), (EPSILON, 5)]
